backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart>=0.0.6
langchain>=0.0.335
langchain-community>=0.0.13
dashscope>=1.17.0
langchain-experimental>=0.0.58
chromadb>=0.5.0
pydantic>=2.4.2
python-dotenv>=1.0.0
tiktoken>=0.5.1
pypdf>=5.2.0
langgraph
langgraph-cli[inmem]